devcontainer の定義を tools.go から外だし。
